+++ tests/functional/splash_ui/test_register_user_ui.py
@@ -8,7 +8,7 @@ from selenium.webdriver.common.by import By
 # from webdriver_manager.chrome import ChromeDriverManager
 
 import tests.functional.constants as const
-from tests.functional.utils import (
+from tests.functional.utils_for_test import (
     click_and_wait,
     find_element_by_css_selector,
     send_keys_to_input_field,
